refactor: Rename project from Windsurf to NovaPilot
This commit renames the project from "Windsurf" to "NovaPilot" across the entire codebase and documentation.

Key changes include:

Directory Renaming:
- `windsurf_core` directory renamed to `novapilot_core`.
- `windsurf_agents` directory renamed to `novapilot_agents`.

Python Code Updates:
- All Python import statements updated to use the new `novapilot_core` and `novapilot_agents` paths.
- Comments in `novapilot_agents/__init__.py` updated.
- In `CodeGenerationAgent`:
    - The "hello world" capability description updated to "Generates a 'Hello, NovaPilot!' Python function."
    - The corresponding generated code string in `_process_task` updated to print "Hello, NovaPilot!".

Documentation Updates:
- Main `README.md`: Project title and all references to "Windsurf" (including directory paths like `windsurf_core`, `windsurf_agents`) updated to "NovaPilot" and its corresponding paths.
- Agent-specific READMEs (`UserInteractionOrchestratorAgent`, `CodeGenerationAgent`): References to "Windsurf" or `windsurf_core` updated to "NovaPilot" or `novapilot_core`.

Verification:
- I executed the `run_example_interaction.py` script successfully after all changes, confirming that imports, agent discovery, capability-based routing, and task execution function correctly with the new project name.
