+++ novapilot_agents/UserInteractionOrchestratorAgent/orchestrator.py
@@ -1,11 +1,11 @@
-# windsurf_agents/UserInteractionOrchestratorAgent/orchestrator.py
+# novapilot_agents/UserInteractionOrchestratorAgent/orchestrator.py # Corrected path in comment
 import asyncio
 import uuid # For generating unique task IDs
 from typing import Dict, Optional, Any, Callable, List # Add List
 
-from windsurf_core.aci import AgentCommunicationInterface
-from windsurf_core.models import Task, ExecutionResult, AgentCapability # Add AgentCapability
-from windsurf_core.message_bus import message_bus
+from novapilot_core.aci import AgentCommunicationInterface # Corrected import
+from novapilot_core.models import Task, ExecutionResult, AgentCapability # Corrected import
+from novapilot_core.message_bus import message_bus # Corrected import
 
 class UserInteractionOrchestratorAgent(AgentCommunicationInterface):
     def __init__(self, agent_id: str):
