+++ novapilot_agents/__init__.py
@@ -1,0 +1,3 @@
+# novapilot_agents/__init__.py
+# This file makes Python treat the directory novapilot_agents as a package.
+# You can leave it empty or add package-level imports here if needed later.
