--- windsurf_agents/__init__.py
@@ -1,3 +1,0 @@
-# windsurf_agents/__init__.py
-# This file makes Python treat the directory windsurf_agents as a package.
-# You can leave it empty or add package-level imports here if needed later.
